Comment in an Issue to display file changes during QC

This function allows a user to insert a comment into a ghqc GitHub Issue that displays changes in the version control information for the Issue’s corresponding file. By default, the comment displays both the original and current commits and hashes for the file. These versions are selected by the user. The comment can optionally display the file difference (“diff”) between the current and previous versions. These changes will likely be implementations of QC feedback.

To use this app, first initialize one or more Issues with ghqc_assign_app.

To comment in an Issue:

Optional: filter to the set of Issues within a Milestone.

Select the Issue to be updated.

Optional: provide a contextualizing message about the changes made to the file (e.g. “Implemented QC feedback for line 20”).

Optional: insert the file difference display into the comment, by selecting “Show file difference”. If displaying the file difference, choose to either:

compare the original version with the current version or,

compare a previous version with the current version.

Optional: preview the comment before posting to the Issue.

Post the comment to the Issue.

Usage

ghqc_resolve_app(
  app_name = "ghqc_resolve_app",
  qc_dir = getwd(),
  lib_path = ghqc_libpath(),
  info_path = ghqc_infopath()
)

Arguments

app_name

the name of the app to run in the background

qc_dir

the directory in which the app is run

lib_path

the path to the ghqc package and its dependencies

info_path

the path to the ghqc configuring information

See also